+++ Tests/Stop-RubrikManagedVolumeSnapshot.Tests.ps1
@@ -1,0 +1,45 @@
+Remove-Module -Name 'Rubrik' -ErrorAction 'SilentlyContinue'
+Import-Module -Name './Rubrik/Rubrik.psd1' -Force
+
+foreach ( $privateFunctionFilePath in ( Get-ChildItem -Path './Rubrik/Private' | Where-Object extension -eq '.ps1').FullName  ) {
+    . $privateFunctionFilePath
+}
+
+Describe -Name 'Public/Stop-RubrikManagedVolumeSnapshot' -Tag 'Public', 'Stop-RubrikManagedVolumeSnapshot' -Fixture {
+    #region init
+    $global:rubrikConnection = @{
+        id      = 'test-id'
+        userId  = 'test-userId'
+        token   = 'test-token'
+        server  = 'test-server'
+        header  = @{ 'Authorization' = 'Bearer test-authorization' }
+        time    = (Get-Date)
+        api     = 'v1'
+        version = '4.0.5'
+    }
+    #endregion
+ 
+    Context -Name 'Parameter Validation' {
+        Mock -CommandName Test-RubrikConnection -Verifiable -ModuleName 'Rubrik' -MockWith {}
+        Mock -CommandName Submit-Request -Verifiable -ModuleName 'Rubrik' -MockWith {
+            @{
+                'id'      = 'id1'
+                'slaId'   = 'slaid'
+                'slaName' = 'Gold'
+            }
+        }
+        It -Name 'Request Fulfilled' -Test {
+            (Stop-RubrikManagedVolumeSnapshot -id 'id1' ).id |
+                Should -BeExactly 'id1'
+        }
+
+        It -Name 'Parameter ID must be present' -Test {
+            { Stop-RubrikManagedVolumeSnapshot -Id  } |
+                Should -Throw "Missing an argument for parameter 'id'. Specify a parameter of type 'System.String' and try again."
+        }
+
+        Assert-VerifiableMock
+        Assert-MockCalled -CommandName Test-RubrikConnection -ModuleName 'Rubrik' -Times 1
+        Assert-MockCalled -CommandName Submit-Request -ModuleName 'Rubrik'  -Times 1
+    }
+}